[CORL-2513]: Integrate React Testing Library with initial test (#3884)
* add testing library and related libraries

* also add package-lock

* add first RTL test

* reorganize testHelpers

* set up more customRenders and createContexts

* move jest-dom import and fix linting

* small refactor

* add css supports for tests

* update imports
