+++ src/core/client/account/test/create.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import App from "coral-account/App";
 import { GQLResolver } from "coral-framework/schema";
 import {
+  createTestContext,
   createTestRenderer,
   CreateTestRendererParams,
 } from "coral-framework/testHelpers";
@@ -11,4 +12,10 @@ export default function create(
   params: CreateTestRendererParams<GQLResolver> = {}
 ) {
   return createTestRenderer<GQLResolver>("account", <App />, params);
+}
+
+export function createContext(
+  params: CreateTestRendererParams<GQLResolver> = {}
+) {
+  return createTestContext<GQLResolver>("account", params);
 }
