+++ src/core/client/account/test/customRenderAppWithContext.tsx
@@ -1,0 +1,16 @@
+import { render } from "@testing-library/react";
+import React from "react";
+
+import App from "coral-account/App";
+import {
+  CoralContext,
+  CoralContextProvider,
+} from "coral-framework/lib/bootstrap";
+
+export default function customRenderAppWithContext(context: CoralContext) {
+  render(
+    <CoralContextProvider value={context}>
+      <App />
+    </CoralContextProvider>
+  );
+}
