+++ src/core/client/stream/test/customRenderAppWithContext.tsx
@@ -1,0 +1,16 @@
+import { render } from "@testing-library/react";
+import React from "react";
+
+import {
+  CoralContext,
+  CoralContextProvider,
+} from "coral-framework/lib/bootstrap";
+import AppContainer from "coral-stream/App";
+
+export default function customRenderAppWithContext(context: CoralContext) {
+  render(
+    <CoralContextProvider value={context}>
+      <AppContainer disableListeners />
+    </CoralContextProvider>
+  );
+}
