+++ src/core/client/test/setupTestFramework.ts
@@ -1,3 +1,4 @@
+import "@testing-library/jest-dom";
 import { toHaveNoViolations } from "jest-axe";
 
 import expectAndFail from "./expectAndFail";
@@ -5,6 +6,8 @@ import expectAndFail from "./expectAndFail";
 // Automatically unmock console.
 import "jest-mock-console/dist/setupTestFramework";
 import "./setupConsole";
+
+(global as any).CSS = { supports: jest.fn() };
 
 // Expose a version of expect that will fail tests immediately
 // when assertion fails. This works even inside of an try-catch block
